requirements.txt
apispec==3.3.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Cython==0.29.20
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-mongoengine==0.9.5
flask-swagger-ui==3.25.0
Flask-WTF==0.14.3
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-mongoengine==0.11.0
mongoengine==0.20.0
mongoengine_goodjson==1.1.8
monotonic==1.5
Naked==0.1.31
pycparser==2.20
pymongo==3.10.1
pyOpenSSL==19.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
